LayoutTests/accessibility/button-with-aria-haspopup-role-expected.txt

This tests the platform role exposed for buttons with aria-haspopup

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


test1 AXRole: AXButton for aria-haspopup=(null)
test2 AXRole: AXPopUpButton for aria-haspopup='true'
test3 AXRole: AXButton for aria-haspopup='false'
test4 AXRole: AXPopUpButton for aria-haspopup='dialog'
test5 AXRole: AXPopUpButton for aria-haspopup='grid'
test6 AXRole: AXPopUpButton for aria-haspopup='listbox'
test7 AXRole: AXPopUpButton for aria-haspopup='menu'
test8 AXRole: AXPopUpButton for aria-haspopup='tree'
test9 AXRole: AXButton for aria-haspopup='foo'
test10 AXRole: AXButton for aria-haspopup=''
PASS successfullyParsed is true

TEST COMPLETE
